services/requirements.txt
tensorflow
numpy
Flask
Flask-Cors
pillow